# ai-native-methodology/examples/poc-14-fsim-corroboration/input/swagger.yaml
openapi: 3.0.3
info:
  title: Todo API
  version: 0.1.0
  description: |
    Small Todo REST API — 외부 사용자 사이드 프로젝트.
    의도된 결함 포함 (production 직전 외부 컨설팅 대상).
servers:
  - url: http://localhost:8000
    description: Local development

paths:
  /users:
    post:
      summary: 신규 사용자 등록
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error (password < 8자)
        '409':
          description: Email already exists (현 구현 미수행 — antipattern AP-FSIM-DATA-001)

  /users/login:
    post:
      summary: 로그인
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '401':
          description: Invalid credentials

  /todos:
    get:
      summary: 본인 Todo 목록 조회
      operationId: listTodos
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
    post:
      summary: Todo 생성
      operationId: createTodo
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'

  /todos/{id}:
    put:
      summary: Todo 수정
      operationId: updateTodo
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        '200':
          description: OK
        '403':
          description: Not the owner
        '404':
          description: Not found
    delete:
      summary: Todo 삭제
      operationId: deleteTodo
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: No content
        '403':
          description: Not the owner
        '404':
          description: Not found

components:
  schemas:
    UserCreate:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
    AuthToken:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: bearer }
    TodoCreate:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        body: { type: string, nullable: true }
        due_date: { type: string, format: date-time, nullable: true }
    TodoUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        body: { type: string, nullable: true }
        due_date: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [pending, done] }
    Todo:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        title: { type: string }
        body: { type: string, nullable: true }
        due_date: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [pending, done] }
        created_at: { type: string, format: date-time }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT